LUA improvements and collision stuff
Lots of hardening - Scripting interface no longer crashes program on fail (resumes if possible, otherwise exits gracefully and prints an error message if DEBUGMODE flag is set).

Terrain now managed independently from other GOs, add it with "attachTerrain" Scene manager command.

Almost every function that can be called from lua will not cause the program to crash at this point. Haven't throughly tested this, but the major issues are accounted for. Try unccommenting all the load functions and running the program, you'll see for yourself.

Currently the bounding boxes for all GOs are drawn, this is temporary to make writing collision easier.

+++ src/Terrain.cpp
@@ -28,9 +28,9 @@ void terrain::setScale(vec3 & toset) {
 
 void terrain::genPlaneIndicies(unsigned size) {
 
-	for (int i = 0; i < size - 1; i++) {
+	for (int i = 0; i < (int) size - 1; i++) {
 		if (i % 2 == 0) {
-			for (int j = 0; j < size; j++) {
+			for (int j = 0; j < (int) size; j++) {
 				planInd.push_back(j + i * size);
 				planInd.push_back(j + (i + 1) * size);
 			}
